refactor: Move `isLocalyInstalled` util to CLI context
