+++ lib/cli/handle-error.js
@@ -14,7 +14,7 @@ const { logWarning } = require('../classes/Error');
 const isStandaloneExecutable = require('../utils/isStandaloneExecutable');
 const tokenizeException = require('../utils/tokenize-exception');
 const logDeprecation = require('../utils/logDeprecation');
-const resolveIsLocallyInstalled = require('../utils/is-locally-installed');
+const resolveIsLocallyInstalled = require('./is-locally-installed');
 const isTelemetryDisabled = require('../utils/telemetry/areDisabled');
 const { storeLocally: storeTelemetryLocally, send: sendTelemetry } = require('../utils/telemetry');
 const generateTelemetryPayload = require('../utils/telemetry/generatePayload');
