+++ lib/cli/is-locally-installed.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const path = require('path');
-const localServerlessPath = require('../cli/local-serverless-path');
+const localServerlessPath = require('./local-serverless-path');
 
 const serverlessPath = path.resolve(__dirname, '../..');
