+++ lib/utils/telemetry/generatePayload.js
@@ -12,7 +12,7 @@ const { getConfigurationValidationResult } = require('../../classes/ConfigSchema
 const { triggeredDeprecations } = require('../logDeprecation');
 const isNpmGlobal = require('../npmPackage/isGlobal');
 const localServerlessPath = require('../../cli/local-serverless-path');
-const resolveIsLocallyInstalled = require('../../utils/is-locally-installed');
+const resolveIsLocallyInstalled = require('../../cli/is-locally-installed');
 const getRequire = require('../../utils/get-require');
 const ci = require('ci-info');
 const AWS = require('aws-sdk');
